Initialize repository
Created repository to house the code for the contact angles script. As
this code base had begun to deviate significantly from the original, I
decided that it would be better to give it its own repository, rather
than leave it tied to the original. This should provide new users a
fresh place to start, with easier logic to follow through the process.

In updating to this new repository, I have also included several
improvements to the analysis script, which should now allow users on
Linux systems to pipe command line arguments to the Python calls. In
addition, I have added a Windows script which should allow the same
types of batch analysis via the command line.