Include fitting with Bashforth-Adams model
Rather than limit the analysis to only two fitting types (linear and
spherical cap approximation), it was suggested that we include the
Bashforth-Adams model as well, which should take into account the
effects of gravity on droplet shape. In this way, even if the droplet
volume is large or has contact angle > 150 deg, the fitting should
remain accurate near the baseline, and shouldn't be as sensitive
to changes as might be the case with the linear or spherical cap.
Even though we include this fit type, it is the most computationally
expensive and should be used when necessary. The linear fitting is
the least costly but is best suited for classification rather than
strict quantitative analysis, while the spherical cap approximation
is a reasonable hybrid of the two extremes.

With the original Bashforth-Adams fit that was implemented, I non-
dimensionalized to the droplet height, which resulted in an un-
inentional normalization of all of the Bond numbers to the same
value regardless of droplet, which made the BA fit nearly identical
to the circular fit. As a result, the fit quality was poor for droplets
sagging under the effects of gravity. This has now been remedied by
fitting to the capillary length (units of px) and droplet radius of
curvature at the apex (units of px). If a user wants to extract more
information about the droplet - such as a dimensional value for these
parameters, they will need some scale for the image to convert px to
mm, for example.