Parallelize code and add autocrop functionality
With the Bashforth-adams fitting methodology, the fits were fairly
slow given the computational expense. To address this issue, I switched
the computation from serial on each frame to be analyzed to analyzing
in parallel across multiple cores available. In addition to the JIT
compilation done by numba, this has afforded decent speedups in
processing time.

With everything else being automated, the cropping was the one part of
code that still required user intervention, which meant that McLain
had to sit at the computer for long intervals just to click a rectangle
every so often. Instead, this new version implements an auto-crop
strategy that computes the histogram of points and looks for changes
in the density of edge points that have been extracted. In this way
we can decently estimate where the droplet is within the image, speeding
up the processing significantly from the user's perspective.