Implement Hough Transform Autocropping
Rather than relying on density of points for auto-cropping routine,
we can take advantage of the fact that the droplet is the "most
circular" component in the image. As a result, the CHT can readily
be used to locate the desired crop region. In addition, we can use
the fact that the baseline will be the most "linear" portion of the
image to identify and fit the baseline, without requiring δ_baseline

This commit also (bad practice) incorporates a new interface for users
who want to run droppy through interactive sessions. This can be
accessed through droppy.run() after import droppy. See the docs for
more.